# docs/swagger.yaml
definitions:
  handler.AddMemberReq:
    properties:
      email:
        type: string
      role:
        enum:
        - editor
        - viewer
        type: string
    required:
    - email
    - role
    type: object
...
  handler.CreateTaskReq:
    properties:
      collaboration_mode:
        type: string
      content_md:
        type: string
      doc_type:
        type: string
      due_at:
        type: string
      priority:
        type: integer
      project_id:
        type: integer
      status:
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - project_id
    - title
    type: object
  handler.DueCallbackReq:
    properties:
      task_id:
        minimum: 1
        type: integer
      triggered_at:
        type: string
    required:
    - task_id
    type: object
...
  handler.RemoveMemberReq:
    properties:
      user_id:
        type: integer
    required:
    - user_id
    type: object
  handler.SaveDocumentContentReq:
    properties:
      content_md:
        type: string
      expected_version:
        type: integer
    required:
    - content_md
    - expected_version
    type: object
...
  handler.UpdateTaskReq:
    properties:
      clear_due_at:
        type: boolean
      content_md:
        type: string
      due_at:
        type: string
      expected_version:
        type: integer
      priority:
        type: integer
      project_id:
        type: integer
      sort_order:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
...
  models.Task:
    properties:
      collaboration_mode:
        type: string
      content_md:
        type: string
      created_at:
        type: string
      doc_type:
        type: string
      due_at:
        type: string
      id:
        type: integer
      members:
        items:
          $ref: '#/definitions/models.TaskMemberInfo'
        type: array
      notified:
        type: boolean
      priority:
        type: integer
      project_id:
        type: integer
      sort_order:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      version:
        type: integer
    type: object
...
  models.TaskMemberInfo:
    properties:
      joined_at:
        type: string
      role:
        type: string
      user:
        $ref: '#/definitions/models.UserInfo'
      user_id:
        type: integer
    type: object
...
  models.UserInfo:
    properties:
      avatar_url:
        type: string
      username:
        type: string
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  response.Resp:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.TaskListResult:
    properties:
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      total:
        format: int64
        type: integer
    type: object
info:
  contact: {}
paths:
  /api/internal/scheduler/task-due:
    post:
      consumes:
      - application/json
      description: 内部调度器回调接口，用于触发任务到期通知
      parameters:
      - description: 调度令牌
        in: header
        name: X-Scheduler-Token
        required: true
        type: string
      - description: 回调信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/handler.DueCallbackReq'
      produces:
      - application/json
      responses:
        "200":
          description: 回调处理成功
          schema:
            $ref: '#/definitions/response.Resp'
        "401":
          description: 令牌无效
          schema:
            $ref: '#/definitions/response.Resp'
      summary: 任务到期回调
      tags:
      - Internal
...
  /documents/{id}/content:
    patch:
      consumes:
      - application/json
      description: Saves plain Markdown content for diary documents. Only the owner
        can save, and this does not write the Yjs update log.
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      - description: Content payload
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/handler.SaveDocumentContentReq'
      produces:
      - application/json
      responses:
        "200":
          description: Saved
          schema:
            allOf:
            - $ref: '#/definitions/response.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/response.Resp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Resp'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/response.Resp'
      security:
      - BearerAuth: []
      summary: Save plain Markdown document content
      tags:
      - Document
...
  /projects/{id}/tasks/{task_id}:
    patch:
      consumes:
      - application/json
      description: 更新任务详情（标题、内容、状态、优先级等）
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      - description: 任务ID
        in: path
        name: task_id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateTaskReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/response.Resp'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Resp'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.Resp'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/response.Resp'
      security:
      - BearerAuth: []
      summary: 更新任务
      tags:
      - Task
  /projects/{id}/tasks/{task_id}/members:
    delete:
      consumes:
      - application/json
      description: 移除任务成员
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      - description: 任务ID
        in: path
        name: task_id
        required: true
        type: integer
      - description: 成员ID
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/handler.RemoveMemberReq'
      produces:
      - application/json
      responses:
        "200":
          description: 移除成功
          schema:
            $ref: '#/definitions/response.Resp'
      security:
      - BearerAuth: []
      summary: 移除任务成员
      tags:
      - Task
    post:
      consumes:
      - application/json
      description: 邀请用户参与任务
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      - description: 任务ID
        in: path
        name: task_id
        required: true
        type: integer
      - description: 成员信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/handler.AddMemberReq'
      produces:
      - application/json
      responses:
        "200":
          description: 添加成功
          schema:
            $ref: '#/definitions/response.Resp'
      security:
      - BearerAuth: []
      summary: 添加任务成员
      tags:
      - Task
...
  /tasks:
    get:
      consumes:
      - application/json
      description: 获取指定项目的任务列表
      parameters:
      - description: 项目ID
        in: query
        name: project_id
        required: true
        type: integer
      - description: 任务状态 (todo/done)
        in: query
        name: status
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Resp'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
              type: object
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.Resp'
      security:
      - BearerAuth: []
      summary: 任务列表
      tags:
      - Task
    post:
      consumes:
      - application/json
      description: 在指定项目中创建新任务
      parameters:
      - description: 任务信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTaskReq'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Resp'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.Resp'
      security:
      - BearerAuth: []
      summary: 创建任务
      tags:
      - Task
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定任务
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Resp'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.Resp'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/response.Resp'
      security:
      - BearerAuth: []
      summary: 删除任务
      tags:
      - Task
    get:
      consumes:
      - application/json
      description: 获取单个任务的详细信息
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.Resp'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/response.Resp'
      security:
      - BearerAuth: []
      summary: 查询任务详情
      tags:
      - Task
...
  /tasks/me:
    get:
      consumes:
      - application/json
      description: 获取当前用户参与的所有任务
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Resp'
            - properties:
                data:
                  $ref: '#/definitions/service.TaskListResult'
              type: object
      security:
      - BearerAuth: []
      summary: 查看我参与的任务
      tags:
      - Task
...
swagger: "2.0"